% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphacone.R
\name{alphacone}
\alias{alphacone}
\title{Alpha Diversity Metric Distributions}
\usage{
alphacone(
  x,
  rep = 1000,
  steps = seq(from = 0.001, to = 1, by = 0.01),
  diversity = "shannon",
  lower.q = 0.025,
  upper.q = 0.975,
  replace = FALSE
)
}
\arguments{
\item{x}{The phyloseq object.}

\item{rep}{The number of replications of rarefying to perform.}

\item{steps}{The library sizes to rarefy to as a function of the library sizes of samples.}

\item{diversity}{The diversity metric. By default, the Shannon Index will be utilized in the generation of the dataframe. Diversity indexes available in vegan are supported.}

\item{lower.q}{The lower quantile for the distribution cone. By default it will provide the 2.5th.}

\item{upper.q}{The upper quantile for the distribution cone. By default, the 97.5th will be applied.}

\item{replace}{Whether to use replacement during sampling.}
}
\value{
A dataframe of the distribution of the diversity metric.
}
\description{
alphacone() will generate a dataframe of the alpha diversity metric for different library sizes. Different rarefied library sizes may have impact on both the measured value of the diversity metric and the perceived variation in the metric values. Generations of the distribution of the alpha diversity metric allows for a comprehensive examination of the alpha diversity metric values to account for variation introduced into the diversity metric as an artifact of rarefied library size.
}
\examples{
library(mirlyn)
data(example)

\dontrun{
alphacone_example <- alphacone(example, rep = 100)
}


}
