% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphawich.R
\name{alphadivDF}
\alias{alphadivDF}
\title{Alpha Diversity Dataframe}
\usage{
alphadivDF(x, diversity = "shannon")
}
\arguments{
\item{x}{The mirl object.}

\item{diversity}{Diversity index to be applied. By default, the Shannon Index will be utilized in the generation of the dataframe. Diversity indexes available in vegan are supported.}
}
\value{
A dataframe consisting of sample metadata and selected diversity indexes.
}
\description{
alphadivDF() will generate a dataframe of alpha-diversity values from the mirl object allowing for characterization of uncertainty and variation introduced through random subsampling during rarefying.
}
\examples{
library(mirlyn)
data(example)

\dontrun{
mirlexample <- mirl(example, rep = 100)
alphadiv_df <- alphadivDF(mirlexample)
}

}
