% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphawich.R
\name{alphawichVis}
\alias{alphawichVis}
\title{This is the alphawhichvis function.}
\usage{
alphawichVis(alphawichDF, yvar = "DiversityIndex", xvar, colorvar = "Group")
}
\arguments{
\item{alphawichDF}{The alphawichDF object.}

\item{yvar}{The yvar column name.}

\item{xvar}{The xvar column name.}
}
\value{
A ggplot object.
}
\description{
The description.
}
\examples{
# This is one way of using it.
#library(mirlyn)
#data(exampleobject)

#alphawichvis(exampleobject)

}
