% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asvrenamer.R
\name{asv_rename}
\alias{asv_rename}
\title{Assign ASV identifiers to sequence variants}
\usage{
asv_rename(x, file = file.path(getwd(), "ASVids.tsv"))
}
\arguments{
\item{x}{The \code{phyloseq} object.}

\item{file}{The path to the tsv generated file of sequence variant identifiers.}
}
\value{
x the modified phyloseq object.
}
\description{
The asv_rename function will assign unique ASV identifiers that are easy to each sequence
variant and will write a tsv file to your working directory for a reference list of all ASV names.
}
\examples{
library(mirlyn)
library(phyloseq)
data(GlobalPatterns)
\dontrun{
asv_rename(GlobalPatterns)
}

}
