% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betamat.R
\name{betamatPCA}
\alias{betamatPCA}
\title{Beta Diversity Matrix - PCA}
\usage{
betamatPCA(x, transformation = "hellinger", dsim = "bray")
}
\arguments{
\item{x}{The mirl() object.}

\item{transformation}{Supported standardization methods in community ecology in decostand() from library(vegan). Hellinger transformations must be applied to ecological data in order to allow for application in Euclidean space.}

\item{dsim}{The dissimilarity metric supported by vegdist() from library(vegan). By default, the Bray-Curtis dissimilarity matrix will be applied.}
}
\value{
A prcomp object
}
\description{
betamatPCA() will perform a PCA on the mirl() object.
}
\examples{
 library(mirlyn)
 data(example)
 mirlexample <- mirl(example, rep = 10)

 betamatPCA_object <- betamatPCA(mirlexample, dsim = "jaccard")


}
