% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divana.R
\name{divana}
\alias{divana}
\title{Diversity Analysis (DivAna)}
\usage{
divana(
  x,
  xvar,
  sample = "SampleID",
  lowabun = 0.01,
  reps = 1000,
  alphadiv = "Shannon",
  taxcols = NULL,
  divcols = NULL,
  betatransform = "hellinger",
  betadist = "bray"
)
}
\arguments{
\item{x}{The phyloseq object.}

\item{xvar}{The independent categorical variable of interest.}

\item{sample}{The sample identifier.}

\item{lowabun}{The cutoff threshold for low abundance taxonomic clustering. By default, it will cluster all taxonomic groups present at less than 1\% abundance into a low abundance group. Set to NULL if do not want any low abundance clustering.}

\item{reps}{The number of replicates to be included for rarefying. By default, it will conduct 1000 replicates.}

\item{alphadiv}{The alpha diversity metric. By default, the Shannon Index will be utilized in the generation of the dataframe. Diversity indexes available in vegan are supported.}

\item{taxcols}{A colour vector for taxonomic composition bar charts. Must include enough colours for all taxonomic groups present at all 7 levels.}

\item{divcols}{A colour vector for diversity analyses plots.}

\item{betatransform}{Supported standardization methods in community ecology in decostand() from library(vegan). Hellinger transformations must be applied to ecological data in order to allow for application in Euclidean space.}

\item{betadist}{The dissimilarity metric supported by vegdist() from library(vegan). By default, the Bray-Curtis dissimilarity matrix will be applied.}
}
\value{
A list of ggplot objects and a dataframe for alpha-diversity index values.
}
\description{
divana() will conduct a complete analysis from taxonomic composition bar charts to beta-diversity ordinations on samples. It includes the multiple iterations of rarefying and the generation of all associated plots with taxonomic composition, rarefaction curves, alpha diversity and beta-diversity analyses.
}
\examples{
library(mirlyn)
data(example)

complete_analysis <- divana(example, "time")

}
