% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example}
\alias{example}
\title{An example \code{phyloseq} object.}
\format{
A \link[phyloseq:phyloseq-class]{phyloseq::phyloseq} object.
}
\usage{
example
}
\description{
A \code{phyloseq} object with six samples.
}
\keyword{datasets}
