% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asvrenamer.R
\name{fasta_rename}
\alias{fasta_rename}
\title{Assign ASV identifiers to FASTA}
\usage{
fasta_rename(fasta, fixed = file.path(getwd(), "ASVids.tsv"))
}
\arguments{
\item{fasta}{The FASTA file to assign ASV identifiers to.}

\item{fixed}{The ASV reference list created by \code{\link[=asv_rename]{asv_rename()}}.}
}
\value{
\code{NULL}. The FASTA file is modified in place.
}
\description{
The filt_fasta_rename function will assign corresponding ASV identifiers to a FASTA file for use in other applications.
}
