% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asvrenamer.R
\name{filt_fasta_rename}
\alias{filt_fasta_rename}
\title{Assign ASV identifiers to FASTA}
\usage{
filt_fasta_rename(filename, fixed)
}
\arguments{
\item{filename}{The FASTA file to assign ASV identifiers to.}

\item{fixed}{The ASV identifiers.}
}
\value{
FASTA file with modified ASV identifiers.
}
\description{
The filt_fasta_rename function will assign corresponding ASV identifiers to a FASTA file for use in other applications.
}
