% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartax.R
\name{fullbartax}
\alias{fullbartax}
\title{Taxonomic Composition Stacked Barcharts - 7 levels}
\usage{
fullbartax(x, xvar = "Sample", lowabun = 0.01, yvar = "Abundance", cols = NULL)
}
\arguments{
\item{x}{is a phyloseq object.}

\item{xvar}{is a categorical variable selected from the metadata.}

\item{lowabun}{is the cut off value for a low-abundant group.}

\item{yvar}{is the relative abundance of reads.}

\item{cols}{is a vector containing specified colour values (sufficient number for number of taxonomic groups found at all levels).}
}
\description{
The fullbartax() function will create taxonomic composition for barcharts for all 7 taxonomic levels.
}
\examples{
\dontrun{
library(mirlyn)
data(example)
fullbartax(example,"Sample", lowabun = 0.01, yvar="Abundance")
}

}
