% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asv_table.R
\name{get_asv_table}
\alias{get_asv_table}
\title{Compiled ASV Table with Abundance & Taxonomic Classification}
\usage{
get_asv_table(
  x,
  OTU = "OTU",
  Sample = "sample",
  Abundance = "abundance",
  Kingdom = "Kingdom",
  Phylum = "Phylum",
  Class = "Class",
  Order = "Order",
  Family = "Family",
  Genus = "Genus",
  Species = "Species",
  proportion = TRUE
)
}
\arguments{
\item{x}{The object from \code{\link[=phyloseq_to_df]{phyloseq_to_df()}}.}

\item{OTU}{The ASV column name. By default, the OTU column name is "OTU".}

\item{Sample}{The Sample ID column name. By default, the Sample ID column name is
"sample" from the \code{\link[=phyloseq_to_df]{phyloseq_to_df()}} function.}

\item{Abundance}{The Abundance column name. By default, the Sample ID column name
is "abundance" from the \code{\link[=phyloseq_to_df]{phyloseq_to_df()}} function.}

\item{Kingdom}{The taxonomic Kingdom column name. By default, the taxonomic Kingdom
column name is "Kingdom" from the \code{\link[=phyloseq_to_df]{phyloseq_to_df()}} function.}

\item{Phylum}{The taxonomic Phylum column name. By default, the taxonomic Phylum column
name is "Classified.Phylum" from the \code{\link[=phyloseq_to_df]{phyloseq_to_df()}} function.}

\item{Class}{The taxonomic Class column name. By default, the taxonomic Class column
name is "Classified.Class" from the \code{\link[=phyloseq_to_df]{phyloseq_to_df()}} function.}

\item{Order}{The taxonomic Order column name. By default, the taxonomic Order column
name is "Classified.Order" from the \code{\link[=phyloseq_to_df]{phyloseq_to_df()}} function.}

\item{Family}{The taxonomic Family column name. By default, the taxonomic Family column
name is "Classified.Family" from the \code{\link[=phyloseq_to_df]{phyloseq_to_df()}} function.}

\item{Genus}{The taxonomic Genus column name. By default, the taxonomic Genus
column name is "Genus" from the \code{\link[=phyloseq_to_df]{phyloseq_to_df()}} function.}

\item{Species}{The taxonomic Species column name. By default, the taxonomic Species
column name is "Classified.Species" from the \code{\link[=phyloseq_to_df]{phyloseq_to_df()}} function.}

\item{proportion}{A Boolean value determining whether abundances of ASV should be
expressed as raw read counts or \% abundance of the library. By default,
the \% abundance will be applied.}
}
\value{
A dataframe.
}
\description{
\code{\link[=get_asv_table]{get_asv_table()}} will produce a compiled ASV table with abundance and taxonomic
classifications using the output from \code{\link[=phyloseq_to_df]{phyloseq_to_df()}}.
In combination with the \code{\link[dplyr:filter]{dplyr::filter()}} function, users may select a
specific taxonomic group.
}
\examples{
library(mirlyn)
data(example)
example_df <- phyloseq_to_df(example)

# ASV Table for All ASV
example_asv_table <- get_asv_table(example_df, OTU = "Row.names")

# Filter by Taxonomic Level of Interest
library(dplyr)
cyano_example <- example_df \%>\% filter(Phylum == "Cyanobacteria")
cyano_example_asv_table <- get_asv_table(cyano_example, OTU = "Row.names")

}
