% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heat.R
\name{plot_heat}
\alias{plot_heat}
\title{Heatmaps for Taxonomic Group of Interest}
\usage{
plot_heat(x, taxlevel, taxaname, xvar, yvar, fillvar)
}
\arguments{
\item{x}{is a \code{data.frame} from \code{\link[=phyloseq_to_df]{phyloseq_to_df()}}.}

\item{taxlevel}{is the taxonomic level of the taxa you're interested in. By
default, it will look at the phylum level.}

\item{taxaname}{is the taxonomic name of interest.}

\item{xvar}{is a categorical variable selected from the metadata.}

\item{yvar}{is a secondary categorical variable selected from the metadata.}

\item{fillvar}{is the column name of the abundance.}
}
\value{
A \code{ggplot} object.
}
\description{
The \code{\link[=plot_heat]{plot_heat()}} function will create heatmaps visualizing the relative abundance of
a taxonomic group of interest from a \code{data.frame}.
}
\examples{
library(mirlyn)
library(dplyr)
data(example)

example_df <- phyloseq_to_df(example)
example_df_phylum <- example_df \%>\%
    group_by(sample, Id, Phylum) \%>\%
    summarise(abundance = sum(abundance))\%>\%
    mutate(Proportion = abundance/sum(abundance)*100)
plot_heat(example_df_phylum, taxlevel = "Phylum",taxaname = "Cyanobacteria",
    xvar = "sample", yvar = "Id", fillvar = "Proportion")

}
