% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarecurve.R
\name{rarecurve}
\alias{rarecurve}
\title{Rarefaction Curve}
\usage{
rarecurve(x, sample = "Sample")
}
\arguments{
\item{x}{The rarefy_whole_rep() object.}

\item{sample}{The sample ID which will be used to assign colours to different samples. Leave blank if don't want colours.}
}
\value{
A ggplot object of the rarefaction curve showing the observed ASV count according to library size of each sample.
}
\description{
rarecurve() will generate a visualization of the rarefaction curve showing the associated library size.
}
\examples{
library(mirlyn)
data(example)

\dontrun{
rarefy_whole_rep_ex <- rarefy_whole_rep(example, rep = 100)
rarecuruve_ex <- rarecurve(rarefy_whole_rep_ex, sample = "Sample")
}

}
