% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarecurve.R
\name{rarefy_whole_rep}
\alias{rarefy_whole_rep}
\title{Rarefy Entire Sample Set}
\usage{
rarefy_whole_rep(
  x,
  rep = 100,
  steps = seq(from = 0.05, to = 1, by = 0.05),
  set.seed = NULL,
  mc.cores = 1L
)
}
\arguments{
\item{x}{The phyloseq object.}

\item{rep}{The number of replicates to be performed. More replicates will take longer to conduct but will provide a smoother distribution for analyses.}

\item{steps}{The steps in library sizes to rarefy to.}

\item{set.seed}{The seed value for reproducibility.}
}
\value{
A dataframe containing the observed sequence counts for samples of varying rarefied library sizes.
}
\description{
rarefy_whole_rep() will rarefy each sample for different library sizes repeatedly in order to generate a representative set of data to be used in rarecurve() for generating rarefaction curves.
}
\examples{
library(mirlyn)
data(example)

rarefy_whole_rep_ex <- rarefy_whole_rep(example, rep = 10)

}
