% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphacone.R
\name{alphacone}
\alias{alphacone}
\title{Alpha Diversity Index Distributions}
\usage{
alphacone(
  x,
  rep = 1000,
  steps = seq(from = 0.001, to = 1, by = 0.01),
  diversity = "shannon",
  lower.q = 0.025,
  upper.q = 0.975,
  replace = FALSE
)
}
\arguments{
\item{x}{The phyloseq object.}

\item{rep}{The number of replications of rarefying to perform.}

\item{steps}{The library sizes to rarefy to as a function of the library sizes of samples.}

\item{diversity}{The diversity index. By default, the Shannon Index will be utilized in the generation of the dataframe. Diversity indexes available in vegan are supported.}

\item{lower.q}{The lower quantile for the distribution cone. By default it will provide the 2.5th.}

\item{upper.q}{The upper quantile for the distribution cone. By default, the 97.5th will be applied.}
}
\value{
A ggplot object showing the distribution of the diversity index.
}
\description{
alphacone() will generate a distribution plot of the alpha diversity index for different library sizes. Different rarefied library sizes may have impact on both the measured value of the diversity index and the perceived variation in the index values. Generations of the distribution of the alpha diversity index allows for a comprehensive examination of the alpha diversity index values to account for variation introduced into the diversity index as an artifact of rarefied library size.
}
\examples{
library(mirlyn)
data(example)

alphacone_example <- alphacone(example, rep = 10)


# Don't think this works on multiple samples right now.
}
