% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphawich.R
\name{alphawichVis}
\alias{alphawichVis}
\title{Visualization of Alpha Diversity Indexes}
\usage{
alphawichVis(
  alphawichDF,
  xvar = "Sample",
  yvar = "DiversityIndex",
  colorvar = NULL
)
}
\arguments{
\item{alphawichDF}{The alphawichDF object.}

\item{xvar}{The xvar column name. Users must specify the metadata column to be plotted on the x-axis.}

\item{yvar}{The yvar column name. By default, the diversity index value will be plotted.}

\item{colorvar}{The metadata column name that will be used for colour assignment.}
}
\value{
A ggplot object.
}
\description{
alphawichVis() will visualize the diversity index values generated from alphadivDF().
}
\examples{
# This is one way of using it.
library(mirlyn)
data(example)

\dontrun{
example <- mirl(example, rep = 100)
example <- alphadivDF(example)
alphawichVis(example, "Sample")
}

}
