% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betamat.R
\name{betamatPCA}
\alias{betamatPCA}
\title{Beta Diversity Matrix - PCA}
\usage{
betamatPCA(x, transformation = "hellinger", dsim = "bray")
}
\arguments{
\item{x}{The \code{\link[=mirl]{mirl()}} object.}

\item{transformation}{Supported standardization methods in community ecology in
\code{\link[vegan:decostand]{vegan::decostand()}}. Hellinger transformations must be applied to ecological data
in order to allow for application in Euclidean space.}

\item{dsim}{The dissimilarity metric supported by \code{\link[vegan:vegdist]{vegan::vegdist()}}. By default,
the Bray-Curtis dissimilarity matrix will be applied.}
}
\value{
A \code{prcomp} object.
}
\description{
\code{\link[=betamatPCA]{betamatPCA()}} will perform a PCA on the \code{\link[=mirl]{mirl()}} object.
}
\examples{
library(mirlyn)
data(example)
\dontrun{
mirlexample <- mirl(example, rep = 100)
betamatPCA_object <- betamatPCA(mirlexample, dsim = "jaccard")
}

}
