% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betamat.R
\name{betamatPCAvis}
\alias{betamatPCAvis}
\title{Beta Diversity Visualization}
\usage{
betamatPCAvis(x, geom = "point", groups, reps, colours)
}
\arguments{
\item{x}{The betamatPCAvis object.}

\item{geom}{The geom style to be aplied to the plot.}

\item{groups}{A vector containing the metadata variable being explored.}

\item{reps}{The number of replicates conducted in the initial mirl().}

\item{colours}{A vector containing colour values.}
}
\value{
A ggplot object.
}
\description{
betamatPCAvis() will produce a visualization from the prcomp object generated from betamatPCA_object.
}
\examples{
library(mirlyn)
data(example)
mirlexample <- mirl(example, rep = 10)
betamatPCA_object <- betamatPCA(mirlexample)

betamatPCAvis(betamatPCA_object, groups = c("A", "B", "C", "D","E","F"), reps = 10, colours = c("#000000", "#E69F00", "#0072B2", "#009E73", "#F0E442", "#D55E00"))

}
