% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseqtodf.R
\name{phyloseq_to_df}
\alias{phyloseq_to_df}
\title{Conversion of \code{phyloseq} Object to Dataframe}
\usage{
phyloseq_to_df(x)
}
\arguments{
\item{x}{The \code{phyloseq} object.}
}
\value{
A dataframe object.

#' @examples
library(mirlyn)
data(example)
example_df <- phyloseq_to_df(example)
}
\description{
\code{phyloseq_to_df()} will create a dataframe including OTU abundances, taxonomic classification and metadata extracted from the \code{phyloseq} object.
This dataframe can be exported as a CSV file for easy viewing.
}
